infra/scripts/requirements.txt
azure-search-documents == 11.5.3
azure-identity == 1.24.0
azure-storage-blob == 12.26.0
semantic-kernel==1.35.3
azure-cosmos==4.9.0